session_catalog: carve external_rows.rs (pure move)
External-session row context: id/path resolution, wrapper-index row
merge + application, source labels, resumed-session activity replay —
with tests, verbatim. ExternalSessionContext fields widened (accessed
from mod.rs). Purity-checked; tests conserved at 104.
